refactor cdoe in prep for client support
