+++ lib/model_context_protocol/shared/configuration.rb
