+++ lib/model_context_protocol/shared/content.rb
