+++ lib/model_context_protocol/shared/instrumentation.rb
