+++ lib/model_context_protocol/shared/methods.rb
