+++ lib/model_context_protocol/shared/prompt.rb
@@ -1,6 +1,10 @@
 # typed: strict
 # frozen_string_literal: true
 
+require_relative "prompt/argument"
+require_relative "prompt/message"
+require_relative "prompt/result"
+
 module ModelContextProtocol
   class Prompt
     class << self
