+++ lib/model_context_protocol/shared/prompt/argument.rb
