+++ lib/model_context_protocol/shared/prompt/message.rb
