+++ lib/model_context_protocol/shared/prompt/result.rb
