+++ lib/model_context_protocol/shared/resource.rb
