+++ lib/model_context_protocol/shared/resource_template.rb
