+++ lib/model_context_protocol/shared/string_utils.rb
