+++ lib/model_context_protocol/shared/tool.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative "tool/input_schema"
+
 module ModelContextProtocol
   class Tool
     class << self
