+++ lib/model_context_protocol/shared/tool/annotations.rb
