+++ lib/model_context_protocol/shared/tool/input_schema.rb
