+++ lib/model_context_protocol/shared/tool/response.rb
