+++ lib/model_context_protocol/shared/transport.rb
