+++ lib/model_context_protocol/shared/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module ModelContextProtocol
-  VERSION = "0.7.0"
+  VERSION = "1.0.0-preview"
 end
