+++ model_context_protocol.gemspec
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "lib/model_context_protocol/version"
+require_relative "lib/model_context_protocol/shared/version"
 
 Gem::Specification.new do |spec|
   spec.name          = "model_context_protocol"
@@ -19,9 +19,7 @@ Gem::Specification.new do |spec|
   spec.metadata["homepage_uri"] = spec.homepage
   spec.metadata["source_code_uri"] = spec.homepage
 
-  spec.files = Dir.chdir(File.expand_path("..", __FILE__)) do
-    %x(git ls-files -z).split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-  end
+  spec.files = Dir.glob("lib/**/*.rb").reject { |f| f.match(%r{^(test|spec|features)/}) }
 
   spec.bindir        = "exe"
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
